BACKEND_)/requirements.txt
PyPDF2==3.0.1
spacy==3.7.2
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
tesseract-ocr==0.1.3
pytesseract==0.3.10
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.7
langdetect==1.0.9 